refactor(doctor): TTP readinessをdomainへ抽出する (#3919)
