+++ src/youtube_automation/domains/channel_readiness/__init__.py
@@ -1,0 +1,15 @@
+"""Provider-neutral channel setup and TTP readiness policies."""
+
+from youtube_automation.domains.channel_readiness.readiness import (
+    ReadinessResult,
+    approved_ttp_exceptions,
+    evaluate_initial_setup_readiness,
+    evaluate_ttp_wf_new_readiness,
+)
+
+__all__ = [
+    "ReadinessResult",
+    "approved_ttp_exceptions",
+    "evaluate_initial_setup_readiness",
+    "evaluate_ttp_wf_new_readiness",
+]
